Fix consequences of a sqlite3 adapter bug
the adapter creates tables with a default value not equal null
even if the parameter default => nil is used in migrations or
schema.rb

waiting to find a more elegant solution, a script was created
which drops and recreate the table "feature_type_in_annotations"
giving the correct default value of nil to some of its columns.

This correct the problem that the first time an annotation was
seen, its feature types were by default all hidden.